23/7/25 Version 8
Created a point system that works based on the choices that the user and AI make.
Created a quit and resume function, this means that the user can exit the game at any time and then resume it; coming back to the same round.
Overall quality of life chnages also, removed at empty system.out.println(); lines. Instead replaced them with \n. This makes the code look a lot more cleaner.